data_client/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
redis==4.2.2